Extract Redis operations and station logic into separate modules
